feat: split list.js
